Event Creation
- Command for creation of event drafts
- Command for user reminder offsets configuration
- Button Request system
- Modal Request system
- Constants for discord field validation
- Auto completion for removing offsets
- Fixed entity mapping
- Changed entities to use Long for development QOL
- Removed unused DTOs
- Cleanup job for unconfirmed drafts
- Discord Utility for creating timestamps
- Date helpers

+++ src/main/java/com/th3hero/eventbot/dto/course/Course.java
@@ -3,7 +3,7 @@ package com.th3hero.eventbot.dto.course;
 import lombok.NonNull;
 
 public record Course(
-        @NonNull Integer id,
+        @NonNull Long id,
         @NonNull String code,
         @NonNull String name,
         @NonNull String nickname
