+++ src/main/java/com/th3hero/eventbot/repositories/CourseRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.query.Param;
 import java.util.Optional;
 
 
-public interface CourseRepository extends JpaRepository<CourseJpa, Integer> {
+public interface CourseRepository extends JpaRepository<CourseJpa, Long> {
     Optional<CourseJpa> findCourseJpaByCode(String courseCode);
 }
